requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
frozendict==2.3.8
html5lib==1.1
idna==3.4
importlib-resources==6.0.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.2
mplfinance==0.12.10b0
multitasking==0.0.11
newsapi-python==0.2.7
numpy==1.25.2
packaging==23.1
pandas==2.1.0
peewee==3.17.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==1.26.6
webencodings==0.5.1
yfinance==0.2.33
zipp==3.16.2
